requirements.txt
# Web Framework & Server
fastapi
uvicorn
gunicorn

# Database & Migrations
sqlalchemy
psycopg2-binary
alembic

# Caching
redis

# Utilities & API Calls
python-dotenv
requests
